Update dependencies and Node version in Tailwind-material package
Updated dependencies in the yarn.lock file and package.json to their latest versions, incorporating newer functionalities and bug fixes. Also added @size-limit/esbuild-why as a new dependency for better understanding of the bundle size composition. In the Github workflow, Node 16 was used to run the jobs, ensuring compatibility with the most recent LTS version of Node.js. These updates aim to enhance the performance, compatibility and stability of the package.