Add NPM_TOKEN to GitHub Actions and update release.config.js
This commit updates two files '.github/workflows/release.yml' and 'release.config.js'. In the 'release.yml', the NPM_TOKEN was added into the environment variables of GitHub Actions. This change allows semantic-release to automatically publish the package to NPM. In 'release.config.js', it was set to release builds from the 'master', 'next', and 'beta' branches. It also added more plugins for semantic-release including changelog, npm, and git. These improvements provide better automatic deployment and versioning for the project.

In the 'package.json', the project’s name was updated from "tailwind-material" to "@udixio/tailwind-material". This helps to increase the visibility and recognizability of the project on NPM by associating it with the organization name.

Lastly, the 'jest.config.js' file was reformatted for cleaner code. The removal of the specific comment helps to eliminate any possible confusion in the future.