feat: Change name and contents of tailwind-material package
Package name changed from "tailwind-material" to "@udixio/tailwind-material". Updated dependencies and peerDependencies, removing the "peer" attribute and modifying the version requirements for "tailwindcss". Also, the majority of text in README.md was replaced with information about the Tailwind Material package. These changes are likely related to tailoring the package for a specific project or purpose, improving compatibility, and providing more relevant information about the package.