fix: Update tests in tailwind-material package
The tests in 'index.test.ts' have been extended to include color assertions for 'primary-light', 'secondary-light' and 'tertiary-light'. These additional lines of code will enable the function to better cross verify if the generated color codes through createMaterialTheme function are accurate or not. This will help in ensuring the intended user experience in terms of consistent UI color theming. However, the test cases for 'primary-dark', 'secondary-dark' and 'tertiary-dark' are currently commented out and not necessary, hence need not be taken into account.