feat: Add variant parameter to materialTheme
The 'materialTheme' function now accepts 'variant' as a parameter to provide flexibility in setting the color theme variant. This change ensures that the chosen variant in the 'createMaterialTheme' function can properly affect the color scheme applied across the application. The default value is set to 'TONAL_SPOT' if the 'variant' is not specified.