fix: Export Variant from material-theme in index.
A Variant parameter has been added to the exports in index.ts file allowing it to be passed as an argument to the 'materialTheme' function. This update provides users more control in customizing the application's color scheme through the 'createMaterialTheme' function. Default value remains 'TONAL_SPOT' when 'variant' isn't explicitly defined.