fix: In the Material Theme class, we made a change to retain the original color of non-primary color schemes when they are not specifically set, instead of defaulting them to the primary color. This addressed a problem where non-primary color schemes defaulted to the primary color if not specifically set.
Further, this change ensures that each color scheme retains its original color, even when it's not explicitly set, avoiding any potential color clashes in the interface. We have also updated our testing to reflect these changes and confirm accuracy. Modifications to tests included the addition of tertiary and neutral colors, and checks for the primary-container-light color. Additionally, comment lines for the variant and contrastLevel were added for future testing considerations.