feat: Updated MaterialFonts, typography styles, and Tailwind Material version
The MaterialFonts is updated to accept font family parameters. The typography styles are completely rewritten for different sizes. Along with this, added and upgraded the Tailwind Material version from 1.0.0-beta.11 to 1.0.0-beta.12 in both package.json and package-lock.json files. The reason for these changes is to improve typography styles, by allowing more customizable font families and having more accurately defined typography styles. This would increase the flexibility for UI development. Updating the package also brings new features and improvements from the latest version.

Additionally, the text classes in button.tsx and tab.tsx are modified to align with the updated typography styles. The changes in tailwind.config.ts and index.scss are also made to ensure the correct use of the updated styles. The removal of utilities in index.scss is due to the new typography styles being handled inside MaterialFonts now.