Refactor variable names for clarity and enforce encapsulation
In file `main.ts`, the variable `builder` was renamed to `themeManager` to better reflect its instantiation from the `ThemeManager` class. This makes the code more self-explanatory.

Additionally, in `font-theme.ts`, the `pixelUnit` variable has been made private for better encapsulation as it was not being accessed outside the class. This change improves the security and structure of our code.