Refactor codebase and update tailwind-material
The tailwind-material package was updated to version 1.0.0-beta.14. Alongside this update, improvements were made to the codebase structure for better clarity and maintainability. Folders and files were renamed for clearer context, like changing 'color' to 'colorPalette' and removing an unused variable "contentAccentToneDelta". Also, a new 'Color' class was created to handle color-related features, which allows for a more structured approach to manage color options. Some utilities from 'colors' were transferred to a new directory 'material-color-utilities'. This update focuses on improving the overall code readability and structure.