Refactor `themeFigma` variable and enhance loop checks in fontTheme.ts
In the fontTheme.ts file, the `const` declaration of the `themeFigma` variable has been replaced with `let` to allow the variable to be re-assigned or mutated. This change is necessary as `themeFigma` variable's properties are modified within the script.

In addition, conditional checks have been added to confirm the existence of [roleName as FontRole] property inside `themeFigma.styles` object before execution continues. This mechanism will prevent runtime errors or incorrect data assignment if the relevant property does not exist.
Since these changes involve modifications of basic structure (const-to-let conversion) and protective conditions, they enhance the robustness of the code and protect against potential bugs or errors.