Refactor color theme and added unit tests
Refactored the getColorPalette() method in the colorTheme.ts file to improve efficiency with less redundant code. Removed an unnecessary assignment and simplified the process of creating the color palette.

Also added a new unit test file called stateTheme.test.ts to test the new color theme for the application. The tests are designed to ensure that all the color palettes are correctly generated and flattened.

Moreover, Created a related stub file named stateTheme.stub.ts for mock data used in the unit tests. It contains predefined color palettes and results required for the tests.